Refactor: move the LoginProvider from the dto folder to the utils folder.
